Crate iota_sdk

The IOTA Rust SDK

It aims at providing a similar SDK functionality like the one existing for TypeScript. IOTA Rust SDK builds on top of the JSON RPC API and therefore many of the return types are the ones specified in iota_types.

The API is split in several parts corresponding to different functionalities as following:

CoinReadApi - provides read-only functions to work with the coins

EventApi - provides event related functions functions to

GovernanceApi - provides functionality related to staking

QuorumDriverApi - provides functionality to execute a transaction block and submit it to the fullnode(s)

ReadApi - provides functions for retrieving data about different objects and transactions

TransactionBuilder - provides functions for building transactions

Usage

The main way to interact with the API is through the IotaClientBuilder, which returns an IotaClient object from which the user can access the various APIs.

Getting Started

Add the Rust SDK to the project by running cargo add iota-sdk in the root folder of your Rust project.

The main building block for the IOTA Rust SDK is the IotaClientBuilder, which provides a simple and straightforward way of connecting to an IOTA network and having access to the different available APIs.

Below is a simple example which connects to a running IOTA local network, devnet, and testnet. To successfully run this program, make sure to spin up a local network with a local validator, a fullnode, and a faucet server (see the README for more information).

use iota_sdk::IotaClientBuilder;

#[tokio::main]
async fn main() -> Result<(), anyhow::Error> {
    let iota = IotaClientBuilder::default()
        .build("http://127.0.0.1:9000") // provide the IOTA network URL
        .await?;
    println!("IOTA local network version: {:?}", iota.api_version());

    // local IOTA network, same result as above except using the dedicated function
    let iota_local = IotaClientBuilder::default().build_localnet().await?;
    println!("IOTA local network version: {:?}", iota_local.api_version());

    // IOTA devnet running at `https://api.devnet.iota.cafe`
    let iota_devnet = IotaClientBuilder::default().build_devnet().await?;
    println!("IOTA devnet version: {:?}", iota_devnet.api_version());

    // IOTA testnet running at `https://api.testnet.iota.cafe`
    let iota_testnet = IotaClientBuilder::default().build_testnet().await?;
    println!("IOTA testnet version: {:?}", iota_testnet.api_version());

    // IOTA mainnet running at `https://api.mainnet.iota.cafe`
    let iota_mainnet = IotaClientBuilder::default().build_mainnet().await?;
    println!("IOTA mainnet version: {:?}", iota_mainnet.api_version());

    Ok(())
}

Examples

For detailed examples, please check the APIs docs and the examples folder in the repository.

Re-exports

pub use iota_json as json;

pub use iota_json_rpc_types as rpc_types;

pub use iota_types as types;

Modules

apis

error

iota_client_config

json_rpc_error

wallet_context

Structs

IotaClient

Provides all the necessary abstractions for interacting with the IOTA network.

IotaClientBuilder

Builder for creating an IotaClient for connecting to the IOTA network.

PagedStream

A stream which repeatedly calls an async function which returns a page of data.

Constants

IOTA_COIN_TYPE

IOTA_DEVNET_GAS_URL

IOTA_DEVNET_GRAPHQL_URL

IOTA_DEVNET_URL

IOTA_LOCAL_NETWORK_GAS_URL

IOTA_LOCAL_NETWORK_GRAPHQL_URL

IOTA_LOCAL_NETWORK_URL

IOTA_LOCAL_NETWORK_URL_0

IOTA_MAINNET_GRAPHQL_URL

IOTA_MAINNET_URL

IOTA_TESTNET_GAS_URL

IOTA_TESTNET_GRAPHQL_URL

IOTA_TESTNET_URL

Traits

PagedFn

A helper trait for repeatedly calling an async function which returns pages of data.